requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
jieba>=0.42.1
numpy>=1.20.0
jupyter>=1.0.0
ipython>=7.0.0
pandas>=1.1.0
tabulate>=0.8.0
pytest>=6.0.0
pytest-benchmark>=3.4.0
